requirements.txt
streamlit
streamlit-cookies-controller
google-genai
PyPDF2
